refactor: move chat-related files to chat directory and update layout and metadata ```
