+++ frontend/src/app/chat/layout.tsx
