+++ frontend/src/app/chat/main-layout.tsx
