+++ frontend/src/app/layout.tsx
@@ -1,4 +1,4 @@
-import type { Metadata } from 'next';
+import type { Metadata, Viewport } from 'next';
 import { Inter } from 'next/font/google';
 import './globals.css';
 import { BaseProviders } from '@/providers/BaseProvider';
@@ -6,10 +6,16 @@ import { BaseProviders } from '@/providers/BaseProvider';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: 'Codefox',
+  title: 'Codefox - The best dev project generator',
   description: 'The best dev project generator',
 };
 
+export const viewport: Viewport = {
+  width: 'device-width',
+  initialScale: 1,
+  maximumScale: 1,
+  userScalable: false,
+};
 export default function RootLayout({
   children,
 }: {
@@ -18,7 +24,11 @@ export default function RootLayout({
   return (
     <html lang="en" suppressHydrationWarning>
       <body className={inter.className}>
-        <BaseProviders>{children}</BaseProviders>
+        <BaseProviders>
+          <div className="min-h-screen w-full bg-gray-50 dark:bg-gray-900 transition-colors">
+            {children}
+          </div>
+        </BaseProviders>
       </body>
     </html>
   );
